4.21 done

+++ osa3/blogilista/controllers/blogs.js
@@ -43,6 +43,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        return response.status(400).json({ error: 'blog not found'})
+    }
+    
+    if (!blog.user || blog.user.toString() !== decodedToken.id.toString()) {
+        return response.status(401).json({ error: 'invalid user'})
+    }
+    
     await Blog.findByIdAndRemove(request.params.id)
 
     response.status(204).end()
